disable parmetis inside Trilinos
Trilinos will not add an rpath to the parmetis directory which will
cause a missing dependency after installation. One workaround would be
to specify CMAKE_INSTALL_RPATH_USE_LINK_PATH=ON when configuring
Trilinos, but I don't know if that has other consequences.
Instead, disable this for now, because we are not using it anyways.