remove python from boost config
boost fails to compile with "pyconfig.h: No such file or directory"
because of a broken include path setting inside boost. Instead of
fixing/updating boost (which we can't do because newer boost version are
also not working correctly), just disable the python support inside
boost. We are not using that anyways.