add adolc
Another optional dependency for deal 9.0.0. It turns out that the latest
release 2.6.3 is not new enough to work, so I created a temporary
release tarball on my github. We should switch to the official releases
after 2.6.4 is out.
The optional master version compiled from source requires "autoconf"
installed, so I don't want to use it by default.